requirements.txt
beautifulsoup4==4.5.1
certifi==2016.9.26
cffi==1.9.1
coverage==4.2
cryptography==1.6
flake8==3.2.1
html5lib==0.999999999
idna==2.1
mccabe==0.5.2
pluggy==0.4.0
py==1.4.31
pyasn1==0.1.9
pycodestyle==2.2.0
pycparser==2.17
pyfiglet==0.7.5
pyflakes==1.3.0
pyOpenSSL==16.2.0
pytest==3.0.5
pytest-cov==2.4.0
PyYAML==3.12
requests==2.12.3
six==1.10.0
tox==2.5.0
vcrpy==1.10.3
virtualenv==15.1.0
webencodings==0.5
wrapt==1.10.8
